requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.2
django-nested-admin==3.2.4
django-simple-history==2.8.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
pylint==2.4.4
python-monkey-business==1.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
